fix: EditEvent form moved to main components
